Run GAMA on a XML file

From a XML file containing an experiment plan, send and run it in GAMA and returns one XML file by simulation containing the output.

call_gama(parameter_xml_file, hpc, output_dir = "")

Arguments

parameter_xml_file
path to an XML file containing an experiment.

hpc
numeric, number of cores

output_dir
path to saved the output of gama. If not specified, current working directory will be used. If `output_dir` doesn't exist, it will be created.

Examples

parameter_xml_file <- system.file("models", "sir.xml", package = "gamar")
outfiles <- call_gama(parameter_xml_file, hpc = 1, output_dir = tempdir())

#> Running experiment plan...